+++ src/plugin/engine/engine.cpp
@@ -10,6 +10,7 @@
 
 #include "common_infrastructure/common_errors.hpp"
 #include "common_infrastructure/constants.hpp"
+#include "common_infrastructure/error_reporting.hpp"
 #include "common_infrastructure/sample_library/attribution_requirements.hpp"
 
 #include "clap/ext/timer-support.h"
@@ -301,61 +302,6 @@ StateSnapshot CurrentStateSnapshot(Engine const& engine) {
     return MakeStateSnapshot(engine.processor);
 }
 
-[[maybe_unused]] auto PrintInstrumentId(InstrumentId id) {
-    DynamicArrayBounded<char, 100> result {};
-    switch (id.tag) {
-        case InstrumentType::None: fmt::Append(result, "None"_s); break;
-        case InstrumentType::WaveformSynth:
-            fmt::Append(result, "WaveformSynth: {}"_s, id.Get<WaveformType>());
-            break;
-        case InstrumentType::Sampler:
-            fmt::Append(result,
-                        "Sampler: {}/{}/{}"_s,
-                        id.Get<sample_lib::InstrumentId>().library.author,
-                        id.Get<sample_lib::InstrumentId>().library.name,
-                        id.Get<sample_lib::InstrumentId>().inst_name);
-            break;
-    }
-    return result;
-}
-
-[[maybe_unused]] void AssignDiffDescription(dyn::DynArray auto& diff_desc,
-                                            StateSnapshot const& old_state,
-                                            StateSnapshot const& new_state) {
-    dyn::Clear(diff_desc);
-
-    if (old_state.ir_id != new_state.ir_id) {
-        fmt::Append(diff_desc,
-                    "IR changed, old: {}:{} vs new: {}:{}\n"_s,
-                    old_state.ir_id.HasValue() ? old_state.ir_id.Value().library.name.Items() : "null"_s,
-                    old_state.ir_id.HasValue() ? old_state.ir_id.Value().ir_name.Items() : "null"_s,
-                    new_state.ir_id.HasValue() ? new_state.ir_id.Value().library.name.Items() : "null"_s,
-                    new_state.ir_id.HasValue() ? new_state.ir_id.Value().ir_name.Items() : "null"_s);
-    }
-
-    for (auto layer_index : Range(k_num_layers)) {
-        if (old_state.inst_ids[layer_index] != new_state.inst_ids[layer_index]) {
-            fmt::Append(diff_desc,
-                        "Layer {}: {} vs {}\n"_s,
-                        layer_index,
-                        PrintInstrumentId(old_state.inst_ids[layer_index]),
-                        PrintInstrumentId(new_state.inst_ids[layer_index]));
-        }
-    }
-
-    for (auto param_index : Range(k_num_parameters)) {
-        if (old_state.param_values[param_index] != new_state.param_values[param_index]) {
-            fmt::Append(diff_desc,
-                        "Param {}: {} vs {}\n"_s,
-                        k_param_descriptors[param_index].name,
-                        old_state.param_values[param_index],
-                        new_state.param_values[param_index]);
-        }
-    }
-
-    if (old_state.fx_order != new_state.fx_order) fmt::Append(diff_desc, "FX order changed\n"_s);
-}
-
 bool StateChangedSinceLastSnapshot(Engine& engine) {
     auto current = CurrentStateSnapshot(engine);
     // we don't check the params ccs for changes
@@ -520,6 +466,9 @@ static void OnMainThread(Engine& engine) {
 
     if (engine.update_gui.Exchange(false, RmwMemoryOrder::Relaxed))
         engine.plugin_instance_messages.UpdateGui();
+
+    if (AutosaveNeeded(engine.autosave_state))
+        QueueAutosave(engine.autosave_state, CurrentStateSnapshot(engine));
 }
 
 void Engine::OnProcessorChange(ChangeFlags flags) {
@@ -547,12 +496,14 @@ Engine::Engine(clap_host const& host,
                   },
           })} {
 
-    last_snapshot.state = CurrentStateSnapshot(*this);
-
     for (auto ccs = shared_engine_systems.settings.settings.midi.cc_to_param_mapping; ccs != nullptr;
          ccs = ccs->next)
         for (auto param = ccs->param; param != nullptr; param = param->next)
             processor.param_learned_ccs[ToInt(*ParamIdToIndex(param->id))].Set(ccs->cc_num);
+
+    last_snapshot.state = CurrentStateSnapshot(*this);
+
+    InitAutosaveState(autosave_state, random_seed, last_snapshot.state);
 
     presets_folder_listener_id =
         shared_engine_systems.preset_listing.scanned_folder.listeners.Add([&engine = *this]() {
@@ -613,6 +564,8 @@ static void PluginOnTimer(Engine& engine, clap_id timer_id) {
 static void PluginOnPollThread(Engine& engine) {
     // we want to poll for attribution text updates
     engine.host.request_callback(&engine.host);
+
+    AutosaveToFileIfNeeded(engine.autosave_state, engine.shared_engine_systems.paths);
 }
 
 usize MegabytesUsedBySamples(Engine const& engine) {